Add player input handling
The GameHandler class will contain the rules, winning condition and game
logic of the Mastermind game. The HumanPlayer class is passed an instance
of GameHandler upon creation instead of the other way around because
HumanPlayer is a robust layer that takes in input from the player.

Add abstractions to the HumanPlayer class so that ideally any game
(other than Mastermind) could be put into a GameHandler.